implement update null  (#3094)
* (tae): small refactor

* (tae): small refactor 2

* (tae): small refactor 3

* (tae): update append value

* (tae): update for append null

* (tae): update for update null

* (tae): update for update null

* (tae): update for update null

* (tae): update for stdvec

* (tae): update for update null

* (tae): update for error check

* (tae): update moengine update

* (tae): update api

* (tae): update for UpdateValue

* (tae): add update null ut

* Fix the bug that the update file expires (#102)

* (tae): fix lint error

+++ pkg/vm/engine/tae/index/zonemap.go
@@ -247,9 +247,9 @@ func (zm *ZoneMap) Unmarshal(buf []byte) error {
 		zm.max = types.DecodeFixed[float32](buf[:4])
 		return nil
 	case types.Type_FLOAT64:
-		zm.min = types.DecodeFixed[uint64](buf[:8])
+		zm.min = types.DecodeFixed[float64](buf[:8])
 		buf = buf[8:]
-		zm.max = types.DecodeFixed[uint64](buf[:8])
+		zm.max = types.DecodeFixed[float64](buf[:8])
 		return nil
 	case types.Type_DATE:
 		zm.min = types.DecodeFixed[types.Date](buf[:4])
